Remove more warnings

+++ modules/base/rendering/renderablesphere.cpp
@@ -311,7 +311,7 @@ void RenderableSphere::render(const RenderData& data, RendererTasks&) {
     }
 
     if (_fadeOutThreshold > -1.0) {
-        float distCamera = glm::distance(
+        double distCamera = glm::distance(
             data.camera.positionVec3(),
             data.position.dvec3()
         );
